# Core: Improve Image Annotation Engine for Co-op Translator