--- a/.gitignore
+++ b/.gitignore
@@ -421,3 +421,6 @@ co_op_translator.egg-info/
 
 # test repo
 test_repo
+
+# .DS Store
+.DS_Store