… getting_started/command-reference.md
… getting_started/supported-languages.md
… poetry.lock
… pyproject.toml
… requirements.txt
… src/co_op_translator/fonts/NotoSansArabic-Medium.ttf
… src/co_op_translator/fonts/font_language_mappings.yml