## Purpose

This PR is to to improve the quality of images translated.

## Description

- Removed text distortion from image translation.
- Added paragraph detection and text alignment (support included for RTL languages)
- Improve background colour detection and speed.
- Adds a flag for optional 'fast mode' translation, with faster image generation at a slight cost to image quality and text alignment.
- Adds reshaping for complex script languages when plotting images. Languages supported: `[ar, fa, ur, he]`.

## Related Issue

Fixed an issue where Arabic image translations containing latin characters would render square characters. This is a known problem with the [Noto Sans Arabic Font](https://github.com/notofonts/arabic/issues/37).

## Does this introduce a breaking change?

- No, neat mode is enabled by default. Fast mode flag `-f` is optional.

When developers merge from main and run the server, azd up, or azd deploy, will this produce an error?
If you're not sure, try it out on an old environment.

- [ ] Yes
- [x] No

## Type of change

- [x] Bugfix
- [x] Feature
- [ ] Code style update (e.g., formatting, local variables)
- [ ] Refactoring (no functional or API changes)
- [x] Documentation content changes
- [ ] Other... Please describe:

## Checklist

Before submitting your pull request, please confirm the following:

- [x] **I have thoroughly tested my changes**: I confirm that I have run the code and manually tested all affected areas.
- [x] **All existing tests pass**: I have run all tests and confirmed that nothing is broken.
- [ ] **I have added new tests** (if applicable): I have written tests that cover the new functionality introduced by my code changes.
- [x] **I have followed the Co-op Translators coding conventions**: My code adheres to the style guide and coding conventions outlined in the repository.
- [x] **I have documented my changes** (if applicable): I have updated the documentation to reflect the changes where necessary.

## Additional context

### Performance Comparison
#### `Phi3getstarted.png`
![image](https://github.com/user-attachments/assets/46843f45-9e8e-49d7-ae9d-55c1eb0e164b)

#### `demo.png`
![image](https://github.com/user-attachments/assets/94bf0da2-3a63-4de9-8998-9fe570ab3527)

#### Time taken to plot sample images:
| Image                 | Old Method (s) | New Method (Neat Mode) (s) | New Method (Fast Mode) (s) |
|-----------------------|---------------|----------------------------|----------------------------|
| Phi3getstarted.png   | 0.2977        | 0.1817                     | 0.0417                     |
| demo.png             | 0.8500        | 0.3741                     | 0.1146                     |

#### Reshaping Example (Arabic)
<p align="left">
  <img src="https://github.com/user-attachments/assets/90d0c70a-9fd6-447e-b8e0-62aeb07b3a75" width="200px">
</p>