@skytin1004 2025-03-17
Hi @timothycdc,

It looks great to me! I just have a question -  does `arabic-reshaper` need to be applied in Markdown translation as well? I'm wondering if the existing Markdown translation also requires this for proper RTL text handling.

Looking forward to your thoughts!