@timothycdc 2025-03-17
Hi @skytin1004 , arabic-reshaper and other reshapers only need to be applied to image rendering with MIL – most modern markdown engines will already do this automatically. I forgot to push the code that does this – have just done it.